requirements.txt
streamlit
google-generativeai
duckdb
pandas
plotly
python-dotenv
